eth1 mocked for interop
