+++ packages/lodestar/src/chain/factory/block/body.ts
@@ -9,8 +9,8 @@ import {ZERO_HASH} from "../../../constants";
 import {OpPool} from "../../../opPool";
 import {IEth1Notifier} from "../../../eth1";
 import {IProgressiveMerkleTree} from "../../../util/merkleTree";
-import {bestVoteData} from "./eth1Data";
 import {generateDeposits} from "./deposits";
+import {computeEpochOfSlot} from "../../stateTransition/util";
 
 export async function assembleBody(
   config: IBeaconConfig,
@@ -25,7 +25,7 @@ export async function assembleBody(
     opPool.attesterSlashings.getAll().then(value => value.slice(0, config.params.MAX_ATTESTER_SLASHINGS)),
     opPool.attestations.getAll().then(value => value.slice(0, config.params.MAX_ATTESTATIONS)),
     opPool.voluntaryExits.getAll().then(value => value.slice(0, config.params.MAX_VOLUNTARY_EXITS)),
-    bestVoteData(config, currentState, eth1)
+    eth1.getEth1Data(config, currentState, computeEpochOfSlot(config, currentState.slot))
   ]);
   //requires new eth1 data so it has to be done after above operations
   const deposits = await generateDeposits(config, opPool, currentState, eth1Data, merkleTree);
