+++ packages/lodestar/src/eth1/impl/ethers.ts
@@ -6,14 +6,14 @@ import {EventEmitter} from "events";
 import {Contract, ethers} from "ethers";
 import {Block, Log} from "ethers/providers";
 import {deserialize} from "@chainsafe/ssz";
-import {bytes32, Deposit, Gwei, number64} from "@chainsafe/eth2.0-types";
+import {BeaconState, bytes32, Deposit, Epoch, Eth1Data, Gwei, number64} from "@chainsafe/eth2.0-types";
 import {IBeaconConfig} from "@chainsafe/eth2.0-config";
 import {Eth1EventEmitter, IEth1Notifier} from "../interface";
 import {isValidAddress} from "../../util/address";
 import {DEPOSIT_CONTRACT_TREE_DEPTH} from "../../constants";
 import {ILogger} from "../../logger";
-import {OpPool} from "../../opPool";
 import {IEth1Options} from "../options";
+import {mostFrequent} from "../../util/objects";
 
 export interface EthersEth1Options extends IEth1Options {
   contract?: Contract;
@@ -139,6 +139,41 @@ export class EthersEth1Notifier extends (EventEmitter as { new(): Eth1EventEmitt
     return Buffer.from(depositCountHex.substr(2), 'hex').readUIntLE(0, 6);
   }
 
+  public async getEth1Data(config: IBeaconConfig, state: BeaconState, currentEpoch: Epoch): Promise<Eth1Data> {
+    const [head, latestStateBlock] = await Promise.all([
+      this.getHead(),
+      this.getBlock('0x' + state.eth1Data.blockHash.toString('hex'))
+    ]);
+    const validVotes = await this.filterValidVotes(config, state.eth1DataVotes, head, latestStateBlock);
+
+    if(validVotes.length === 0) {
+      const requiredBlock = head.number - config.params.ETH1_FOLLOW_DISTANCE;
+      const blockHash = (await this.getBlock(requiredBlock)).hash;
+      const [depositCount, depositRoot] = await Promise.all([
+        this.depositCount(blockHash),
+        this.depositRoot(blockHash)
+      ]);
+      return {
+        blockHash: Buffer.from(blockHash.slice(2), 'hex'),
+        depositCount,
+        depositRoot
+      };
+    } else {
+      const frequentVotes = mostFrequent<Eth1Data>(validVotes, config.types.Eth1Data);
+      if(frequentVotes.length === 1) {
+        return frequentVotes[0];
+      } else {
+        const blockNumbers = await Promise.all(
+          frequentVotes.map(
+            (vote) =>
+              this.getBlock('0x' + vote.blockHash.toString('hex')).then(b => b.number)
+          )
+        );
+        return frequentVotes[blockNumbers.indexOf(Math.max(...blockNumbers))];
+      }
+    }
+  }
+
   private async initContract(): Promise<void> {
     const address = this.opts.depositContract.address;
     const abi = this.opts.depositContract.abi;
@@ -190,5 +225,30 @@ export class EthersEth1Notifier extends (EventEmitter as { new(): Eth1EventEmitt
         signature: Buffer.from(signature.slice(2), 'hex'),
       },
     };
+  }
+
+  private async filterValidVotes(
+    config: IBeaconConfig,
+    votes: Eth1Data[],
+    head: Block,
+    latestStateBlock: Block): Promise<Eth1Data[]> {
+    const potentialVotes = [];
+    for(let i = 0; i < votes.length; i++) {
+      const vote = votes[i];
+      const block = await this.getBlock(vote.blockHash.toString('hex'));
+      if(block
+          && (head.number - block.number) >= config.params.ETH1_FOLLOW_DISTANCE
+          && block.number > latestStateBlock.number
+      ) {
+        const [depositCount, depositRoot] = await Promise.all([
+          this.depositCount(vote.blockHash.toString('hex')),
+          this.depositRoot(vote.blockHash.toString('hex'))
+        ]);
+        if(depositRoot.equals(vote.depositRoot) && depositCount === vote.depositCount) {
+          potentialVotes.push(vote);
+        }
+      }
+    }
+    return potentialVotes;
   }
 }
